TIP-860: Move repositories/queries from enrich bundle to the right BC
