+++ src/Akeneo/Channel/Bundle/Doctrine/Repository/InternalApi/CurrencyRepository.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
+namespace Akeneo\Channel\Component\Repository\InternalApi;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
 use Oro\Bundle\PimDataGridBundle\Doctrine\ORM\Repository\DatagridRepositoryInterface;
 
 /**
- * Job instance repository
+ * Currency repository
  *
- * @author    Romain Monceau <romain@akeneo.com>
- * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
+ * @author    Filips Alpe <filips@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class JobInstanceRepository extends EntityRepository implements DatagridRepositoryInterface
+class CurrencyRepository extends EntityRepository implements DatagridRepositoryInterface
 {
     /**
      * @param EntityManager $em
@@ -29,13 +29,7 @@ class JobInstanceRepository extends EntityRepository implements DatagridReposito
      */
     public function createDatagridQueryBuilder()
     {
-        $qb = $this->createQueryBuilder('j');
-        $qb
-            ->addSelect("j.jobName as jobName")
-            ->addSelect(
-                "CONCAT('pim_import_export.status.', j.status) as statusLabel"
-            )
-            ->andWhere('j.type = :jobType');
+        $qb = $this->createQueryBuilder('c');
 
         return $qb;
     }
