+++ src/Akeneo/Pim/Enrichment/Bundle/Cursor/SequentialEditProduct.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Cursor;
+namespace Akeneo\Pim\Enrichment\Bundle\Cursor;
 
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Client;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Cursor\Cursor;
