+++ src/Akeneo/Pim/Enrichment/Bundle/Doctrine/ORM/Counter/CategoryItemsCounterInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Doctrine\Counter;
+namespace Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Counter;
 
 use Akeneo\Tool\Component\Classification\Model\CategoryInterface;
