+++ src/Akeneo/Pim/Enrichment/Bundle/spec/Doctrine/ORM/Counter/CategoryProductsCounterSpec.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace spec\Pim\Bundle\EnrichBundle\Doctrine\Counter;
+namespace spec\Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Counter;
 
 use Akeneo\Tool\Component\Classification\Repository\CategoryRepositoryInterface;
 use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
