+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Repository/InternalApi/JobTrackerRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
+namespace Akeneo\Platform\Bundle\ImportExportBundle\Repository\InternalApi;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
