+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Resources/config/repositories.yml
@@ -1,0 +1,19 @@
+services:
+    pim_enrich.repository.job_execution:
+        class: 'Akeneo\Platform\Bundle\ImportExportBundle\Repository\InternalApi\JobExecutionRepository'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%akeneo_batch.entity.job_execution.class%'
+
+    pim_enrich.repository.job_tracker:
+        class: 'Akeneo\Platform\Bundle\ImportExportBundle\Repository\InternalApi\JobInstanceRepository'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%akeneo_batch.entity.job_execution.class%'
+
+    pim_enrich.repository.job_instance:
+        class: 'Akeneo\Platform\Bundle\ImportExportBundle\Repository\InternalApi\JobTrackerRepository'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%akeneo_batch.entity.job_instance.class%'
+
