+++ src/Oro/Bundle/PimDataGridBundle/Repository/ClientRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
+namespace Oro\Bundle\PimDataGridBundle\Repository;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
