+++ src/Pim/Bundle/EnrichBundle/Resources/config/category_counters.yml
@@ -1,7 +1,7 @@
 parameters:
-    pim_enrich.doctrine.counter.category_items.class:    Pim\Bundle\EnrichBundle\Doctrine\Counter\CategoryItemsCounter
-    pim_enrich.doctrine.counter.category_registry.class: Pim\Bundle\EnrichBundle\Doctrine\Counter\CategoryItemsCounterRegistry
-    pim_enrich.doctrine.counter.category_product.class: Pim\Bundle\EnrichBundle\Doctrine\Counter\CategoryProductsCounter
+    pim_enrich.doctrine.counter.category_items.class:    Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Counter\CategoryItemsCounter
+    pim_enrich.doctrine.counter.category_registry.class: Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Counter\CategoryItemsCounterRegistry
+    pim_enrich.doctrine.counter.category_product.class: Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Counter\CategoryProductsCounter
 
 services:
     pim_enrich.doctrine.counter.category_product:
