+++ src/Pim/Bundle/EnrichBundle/Resources/config/query_builder.yml
@@ -1,5 +1,5 @@
 parameters:
-    pim_enrich.cursor.sequential_edit_cursor.class: Pim\Bundle\EnrichBundle\Cursor\SequentialEditProduct
+    pim_enrich.cursor.sequential_edit_cursor.class: Akeneo\Pim\Enrichment\Bundle\Cursor\SequentialEditProduct
 services:
     pim_enrich.query.product_query_sequential_edit_builder_factory:
         class: '%pim_catalog.query.elasticsearch.product_query_builder_factory.class%'
