--- src/Pim/Bundle/EnrichBundle/Resources/config/repositories.yml
@@ -1,54 +1,0 @@
-parameters:
-    ## Default UI repositories
-    pim_enrich.repository.group.class:
-    pim_enrich.repository.locale.class:                  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\LocaleRepository
-    pim_enrich.repository.channel.class:                 Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\ChannelRepository
-    pim_enrich.repository.currency.class:                Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CurrencyRepository
-    pim_enrich.repository.job_instance.class:            Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobInstanceRepository
-    pim_enrich.repository.job_execution.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobExecutionRepository
-    pim_enrich.repository.job_tracker.class:             Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobTrackerRepository
-    pim_enrich.repository.client.class:                  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\ClientRepository
-
-services:
-    ## Default UI repositories
-    pim_enrich.repository.locale:
-        class: '%pim_enrich.repository.locale.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.locale.class%'
-
-    pim_enrich.repository.channel:
-        class: '%pim_enrich.repository.channel.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.channel.class%'
-
-    pim_enrich.repository.currency:
-        class: '%pim_enrich.repository.currency.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.currency.class%'
-
-    pim_enrich.repository.job_instance:
-        class: '%pim_enrich.repository.job_instance.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%akeneo_batch.entity.job_instance.class%'
-
-    pim_enrich.repository.job_execution:
-        class: '%pim_enrich.repository.job_execution.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%akeneo_batch.entity.job_execution.class%'
-
-    pim_enrich.repository.job_tracker:
-        class: '%pim_enrich.repository.job_tracker.class%'
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '%akeneo_batch.entity.job_execution.class%'
-
-    pim_enrich.repository.client:
-        class: '%pim_enrich.repository.client.class%'
-        arguments:
-            - '@fos_oauth_server.entity_manager'
-            - '%fos_oauth_server.model.client.class%'
